requirements.txt
Django==4.2.5
environs==9.5.0
Pillow==10.0.1
python-telegram-bot==13.15